+++ CHANGES.md
@@ -14,6 +14,10 @@
 
 - analysis.plot2Dnet() 'view' param to select frontal ('xy') or top-down ('xz') view 
 
+- analysis.plotConn() option to plot stacked bar graph and select different pre and post pops
+
+- Made distinction between number of connections and synaptic contacts (eg. can have 5 synsPerconn)
+
 - Removed 'plast' and 'shape' from conns when not used (issue #177)
 
 - Now using noiseFromRandom123() for NetStims() (issue #89)